changed int -> long
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@542 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/argument.c
@@ -13,7 +13,7 @@
  *   #)
  */
 
-int  Argc(){ return ArgCount; }
+long  Argc(){ return ArgCount; }
 
 /* Argv is an external described in "betaenvbody.bet"
  *   Argv: External
@@ -25,4 +25,4 @@ int  Argc(){ return ArgCount; }
  *  #) 
  */
 
-char *Argv(n) int n; { return ArgVector[n-1]; }
+char *Argv(n) long n; { return ArgVector[n-1]; }
